requirements.txt
aead==0.2
boto3==1.7.65
celery==4.2.0
channels-redis==2.2.1
Django==2.0.6
django-allauth==0.36.0
django-compressor==2.2
django-htmlmin==0.10.0
django-storages==1.6.6
djangorestframework==3.8.2
furl==1.0.1
future==0.16.0
jsmin==2.2.2
Pillow==5.0.0
psycopg2==2.7.3.1
pycrypto==2.6.1
Pygments==2.2.0
python-decouple==3.1
redis==2.10.6
